requirements.txt
fastapi
uvicorn
motor
pymongo
pandas
numpy
lightfm
xgboost
scikit-learn
python-dateutil
python-dotenv
apscheduler